Week 11/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic[email]==2.6.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
cryptography==42.0.5
python-dotenv==1.0.1
aiofiles==23.2.1
